requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3
python-multipart==0.0.6
ultralytics
pandas==2.3.3
boto3==1.42.76
pytest==9.0.2
